Node Restructure and fixes.  -Created node search Controller and user controller.  -Fixed the structure.  -Optimize and organization.
